Move login view to CBV
